Cleaning client handlers recv()
